Lighting bugfixes
The way js handles color rollovers is NOT how love handles them. Either
that or I'm really bad at this. (Actually, those aren't mutually
exclusive.)
Was getting crazy random colors when several elements of the random
color for the light was above ~150.
To address I added a clamp for colors to Display. The result is a
tendency toward white for high intensity lights, rather than a seemingly
random color.