First working version of Brogue Generator
It is SUPER slow. Takes about 3 seconds to gen a map on my laptop. It will
probably be quicker on better hardware, but I haven't tested on anything
else, yet.

The output is way more engaging as a level that any of the other
generators, though. I will definitely be working on optimising the
process.

A small demo script:
```lua
    ROT=require 'vendor/rotLove/rotLove'

    function love.load()
        f  =ROT.Display(80, 30)
        brg=ROT.Map.Brogue(f:getWidth(), f:getHeight())
        brg:create(calbak,true)
    end
    function love.draw() f:draw() end
    function calbak(x, y, val) f:write(val==3 and '*' or val==2 and '+' or val==1 and '#' or '.', x, y) end
    local update=false
    function love.update()
        if update then
            update=false
            brg:create(calbak)
        end
    end
    function love.keypressed(key) update=true end
```