Update brogue generator algo
The Brogue algorithm is something along the lines of:
Build the first room (wether it be the standard first floor room, a cave,
or a rectangle.)
Try 600 random wall spots
on each spot try 15 times to build a corridor+room(.8 random chance) or
room(forced on last 215 iterations.)

That reliably takes three seconds in my code. Whether because I'm bad at
this or whatever.

changing that to try 1000 wall spots once each cuts the time down to about
a half second. That's what this commit does.

I'll keep trying to speed up the room and corridor building (those eat up
most of the time), but for now this is a significant improvement and the
level doesn't really suffer for it.