Added option to Cellular to connect open tiles
Inspired by a recent commit to rot.js, but implemented a little
differently.
On connecting, loop through all tiles. When a floor tile is encountered,
add all floor tiles that are not connected to a new zone table and mark
the tiles with a shared unique int >1.
Once all zones are identified, loop through the zones and connect a random
tile in the zone to a random tile that is not marked as 1 and is not
marked with the same int as this zone.
Loop back through all tiles in this zone and mark tiles as 0.
Zones are connected in the same manner as the Rogue generator connects its
rooms.