Fix for the weird behavior of Display:clear
also TextDisplay:clear

clear was doing this weird thing where the last row wouldn't clear when
you ran it with no arguments (as in `disp:clear()`)

There was this other problem where the width and height arguments were
processed differently. Width was processed as if that were the number of
columns you wanted to clear and height was that minus one.

This makes so that (for an 80x24 display) you can clear the whole thing
with `disp:clear(1, 1, 80, 24)`

fixes #23